first tweeks after starting playing again
I stayed away from the game for a year, and some of my binds
(specially related to jumpthrow) are not allowed anymore.

Applyed some fixes, and took the chance to organize
my cfgs and bkps into a Git repo.

There are still a lot of cleanup to do
(even thought I already got rid of some unsed cfgs files)